+++ vector/src/gplay/java/im/vector/riotx/gplay/push/fcm/VectorFirebaseMessagingService.kt
@@ -196,7 +196,6 @@ class VectorFirebaseMessagingService : FirebaseMessagingService() {
         // This ID can and should be used to detect duplicate notification requests.
         val eventId = data["event_id"] ?: return //Just ignore
 
-
         val eventType = data["type"]
         if (eventType == null) {
             //Just add a generic unknown event
@@ -214,10 +213,7 @@ class VectorFirebaseMessagingService : FirebaseMessagingService() {
             )
             notificationDrawerManager.onNotifiableEventReceived(simpleNotifiableEvent)
             notificationDrawerManager.refreshNotificationDrawer()
-
-            return
         } else {
-
             val event = parseEvent(data) ?: return
 
             val notifiableEvent = notifiableEventResolver.resolveEvent(event, session)
@@ -228,8 +224,6 @@ class VectorFirebaseMessagingService : FirebaseMessagingService() {
                     Timber.e("--> ${event}")
                 }
             } else {
-
-
                 if (notifiableEvent is NotifiableMessageEvent) {
                     if (TextUtils.isEmpty(notifiableEvent.senderName)) {
                         notifiableEvent.senderName = data["sender_display_name"]
@@ -246,7 +240,6 @@ class VectorFirebaseMessagingService : FirebaseMessagingService() {
                 notificationDrawerManager.refreshNotificationDrawer()
             }
         }
-
     }
 
     private fun findRoomNameBestEffort(data: Map<String, String>, session: Session?): String? {