+++ vector/src/main/java/im/vector/riotx/features/notifications/SimpleNotifiableEvent.kt
@@ -30,6 +30,7 @@ data class SimpleNotifiableEvent(
         override var isPushGatewayEvent: Boolean = false) : NotifiableEvent {
 
     override var hasBeenDisplayed: Boolean = false
+    override var isRedacted: Boolean = false
     override var lockScreenVisibility = NotificationCompat.VISIBILITY_PUBLIC
 
 }